refactor(infra): switch to snake case
