+++ x-pack/plugins/infra/public/pages/metrics/hosts/hooks/use_tab_id.ts
